implement basic room management functions
Rooms can be created or deleted. The list of currently available rooms is also visible in the client. However, it is not yet possible to join a room.

Bug:
Currently, the room list in the client is updated immediately when a new room is created. To update the room list after deletion, the room must either be deleted twice or a new room must be created.

I have not yet found the cause of this bug.